refactor(web): change directory name
