+++ apps/web/src/pages/find-account/ui/FindIdPage.tsx
@@ -3,7 +3,7 @@ import { AuthHeader } from '~/shared/ui';
 
 export default function FindIdPage() {
   return (
-    <div className="relative flex size-full flex-col">
+    <div className="relative flex size-full flex-col p-16 pt-0">
       <AuthHeader title="아이디 찾기" />
     </div>
   );
